ProjectDAO refactoring, in order to remove from RoadAddressDAO the project related code
